# perf(storage): more simple8b bounds check optimization