Some more minor improvements.

Comparison with 9d5389:
```
name                     old time/op    new time/op    delta
EncodeAll/1_bit-8          25.1µs ± 2%    25.2µs ± 2%    ~     (p=0.754 n=10+10)
EncodeAll/2_bits-8         25.5µs ± 2%    25.3µs ± 1%    ~     (p=0.105 n=10+10)
EncodeAll/3_bits-8         25.6µs ± 2%    25.7µs ± 1%    ~     (p=0.616 n=10+10)
EncodeAll/4_bits-8         25.8µs ± 1%    25.8µs ± 1%    ~     (p=0.968 n=9+10)
EncodeAll/5_bits-8         26.4µs ± 1%    26.6µs ± 2%    ~     (p=0.243 n=9+10)
EncodeAll/6_bits-8         27.1µs ± 1%    27.0µs ± 1%    ~     (p=0.065 n=9+10)
EncodeAll/7_bits-8         27.8µs ± 1%    27.8µs ± 1%    ~     (p=0.796 n=10+10)
EncodeAll/8_bits-8         27.9µs ± 1%    28.1µs ± 3%    ~     (p=0.356 n=9+10)
EncodeAll/10_bits-8        30.5µs ± 8%    28.9µs ± 1%  -5.42%  (p=0.043 n=10+9)
EncodeAll/12_bits-8        31.4µs ± 7%    29.6µs ± 1%  -5.48%  (p=0.000 n=10+9)
EncodeAll/15_bits-8        33.2µs ±14%    31.4µs ± 3%  -5.28%  (p=0.003 n=10+10)
EncodeAll/20_bits-8        34.4µs ± 4%    33.7µs ± 1%  -1.95%  (p=0.004 n=10+10)
EncodeAll/30_bits-8        40.5µs ± 4%    39.1µs ± 1%  -3.60%  (p=0.001 n=10+9)
EncodeAll/60_bits-8        56.3µs ± 2%    55.5µs ± 1%  -1.41%  (p=0.002 n=9+9)
EncodeAll/combination-8     497µs ± 4%     477µs ± 2%  -3.99%  (p=0.000 n=10+9)
Encode-8                   2.55µs ± 1%    2.53µs ± 1%  -0.84%  (p=0.033 n=9+9)
Encoder-8                  2.85µs ± 1%    2.74µs ± 1%  -3.67%  (p=0.000 n=8+9)
Decode-8                    608ns ± 2%     598ns ± 1%  -1.59%  (p=0.001 n=8+10)
Decoder-8                  3.21µs ± 1%    3.20µs ± 1%    ~     (p=0.168 n=9+9)

name                     old speed      new speed      delta
Encode-8                 3.21GB/s ± 1%  3.24GB/s ± 1%  +0.86%  (p=0.031 n=9+9)
Encoder-8                2.88GB/s ± 1%  2.99GB/s ± 1%  +3.82%  (p=0.000 n=8+9)
Decode-8                 13.5GB/s ± 2%  13.7GB/s ± 1%  +1.61%  (p=0.002 n=8+10)
Decoder-8                2.55GB/s ± 1%  2.56GB/s ± 1%    ~     (p=0.161 n=9+9)
```