@e-dard 2019-10-25
```
⇒  benchstat ~/simple8b-old.txt ~/simple8b-tanya-imp.txt | grep Dec
Decode-24                    585ns ± 2%     658ns ± 3%  +12.47%  (p=0.000 n=9+10)
Decoder-24                  4.61µs ± 2%    4.65µs ± 2%     ~     (p=0.305 n=10+10)
Decode-24                    0.00B          0.00B          ~     (all equal)
Decoder-24                   0.00B          0.00B          ~     (all equal)
Decode-24                     0.00           0.00          ~     (all equal)
Decoder-24                    0.00           0.00          ~     (all equal)
Decode-24                 14.0GB/s ± 2%  12.5GB/s ± 4%  -11.10%  (p=0.000 n=9+10)
Decoder-24                1.77GB/s ± 1%  1.76GB/s ± 2%     ~     (p=0.497 n=9+10)
```

We are still about `12%` down on decoding from `master`.